Move Path Parsing into Path.
